# infra/swagger.yaml
openapi: 3.0.0
info:
  title: Sandei API
  version: "1.0"
paths:
  /ia/suggest_lineup:
    post:
      summary: Suggest an optimal lineup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LineupRequest'
      responses:
        '200':
          description: Suggested lineup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineupResponse'
  /ia/suggest_tactics:
    post:
      summary: Suggest match tactics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TacticsRequest'
      responses:
        '200':
          description: Suggested tactics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TacticsResponse'
  /ia/predict_match:
    post:
      summary: Predict match result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchPredictionRequest'
      responses:
        '200':
          description: Match prediction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchPredictionResponse'
  /matches/{matchId}/ratings:
    post:
      summary: Register a player rating
      parameters:
        - in: path
          name: matchId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingRequest'
      responses:
        '201':
          description: Rating created
components:
  schemas:
    LineupRequest:
      type: object
      properties:
        players:
          type: array
          items:
            type: string
        formation:
          type: string
      required:
        - players
        - formation
    LineupResponse:
      type: object
      properties:
        lineup:
          type: string
    TacticsRequest:
      type: object
      properties:
        players:
          type: array
          items:
            type: string
        style:
          type: string
      required:
        - players
    TacticsResponse:
      type: object
      properties:
        tactics:
          type: string
    RatingRequest:
      type: object
      properties:
        playerId:
          type: string
        score:
          type: integer
        comment:
          type: string
      required:
        - playerId
        - score
    MatchPredictionRequest:
      type: object
      properties:
        homeTeam:
          type: array
          items:
            type: string
        awayTeam:
          type: array
          items:
            type: string
      required:
        - homeTeam
        - awayTeam
    MatchPredictionResponse:
      type: object
      properties:
        prediction:
          type: string